+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 -   Fix error when docstring is not present in the function. (Issue #1, PR #4)
 -   Add `pyzmq.utils.win32.allow_interrupt` to the server threads, so they can be stopped on Windows. (Issue #2, PR #6)
 -   Move some of the logging to `DEBUG` level. (Issue #3, PR #7)
+-   Add reserved method names to `Thing` class. (Issue #5, PR #8)
 
 [Full Unreleased Changelog](https://github.com/matpompili/caniusethat/compare/v0.2.3...main)